StateManager: extract storage and code tests to dedicated files, use storage cache on/off switch for storage tests
