+++ packages/statemanager/test/stateManager.account.spec.ts
@@ -1,14 +1,4 @@
-import {
-  Account,
-  Address,
-  KECCAK256_RLP,
-  KECCAK256_RLP_S,
-  toBuffer,
-  unpadBuffer,
-  zeros,
-} from '@ethereumjs/util'
-import { keccak256 } from 'ethereum-cryptography/keccak'
-import { bytesToHex } from 'ethereum-cryptography/utils'
+import { Account, Address, KECCAK256_RLP, KECCAK256_RLP_S, toBuffer } from '@ethereumjs/util'
 import * as tape from 'tape'
 // explicitly import `inherits` to fix karma-typescript issue
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -18,7 +8,7 @@ import { DefaultStateManager } from '../src'
 
 import { createAccount } from './util'
 
-tape('StateManager -> General', (t) => {
+tape('StateManager -> General/Account', (t) => {
   for (const accountCacheOpts of [{ deactivate: false }, { deactivate: true }]) {
     t.test('should instantiate', async (st) => {
       const stateManager = new DefaultStateManager({ accountCacheOpts })
@@ -183,286 +173,6 @@ tape('StateManager -> General', (t) => {
       )
 
       st.end()
-    })
-
-    t.test('should dump storage', async (st) => {
-      const stateManager = new DefaultStateManager({ accountCacheOpts })
-      const address = new Address(Buffer.from('a94f5374fce5edbc8e2a8697c15331677e6ebf0b', 'hex'))
-      const account = createAccount()
-
-      await stateManager.putAccount(address, account)
-
-      const key = toBuffer('0x1234567890123456789012345678901234567890123456789012345678901234')
-      const value = toBuffer('0x0a') // We used this value as its RLP encoding is also 0a
-      await stateManager.putContractStorage(address, key, value)
-
-      const data = await stateManager.dumpStorage(address)
-      const expect = { [bytesToHex(keccak256(key))]: '0a' }
-      st.deepEqual(data, expect, 'should dump storage value')
-
-      st.end()
-    })
-
-    t.test("should validate the key's length when modifying a contract's storage", async (st) => {
-      const stateManager = new DefaultStateManager({ accountCacheOpts })
-      const address = new Address(Buffer.from('a94f5374fce5edbc8e2a8697c15331677e6ebf0b', 'hex'))
-      try {
-        await stateManager.putContractStorage(address, Buffer.alloc(12), toBuffer('0x1231'))
-      } catch (e: any) {
-        st.equal(e.message, 'Storage key must be 32 bytes long')
-        st.end()
-        return
-      }
-
-      st.fail('Should have failed')
-      st.end()
-    })
-
-    t.test("should validate the key's length when reading a contract's storage", async (st) => {
-      const stateManager = new DefaultStateManager({ accountCacheOpts })
-      const address = new Address(Buffer.from('a94f5374fce5edbc8e2a8697c15331677e6ebf0b', 'hex'))
-      try {
-        await stateManager.getContractStorage(address, Buffer.alloc(12))
-      } catch (e: any) {
-        st.equal(e.message, 'Storage key must be 32 bytes long')
-        st.end()
-        return
-      }
-
-      st.fail('Should have failed')
-      st.end()
-    })
-
-    t.test('should store codehashes using a prefix', async (st) => {
-      /*
-        This test is mostly an example of why a code prefix is necessary
-        I an address, we put two storage values. The preimage of the (storage trie) root hash is known
-        This preimage is used as codeHash
-  
-        NOTE: Currently, the only problem which this code prefix fixes, is putting 0x80 as contract code
-        -> This hashes to the empty trie node hash (0x80 = RLP([])), so keccak256(0x80) = empty trie node hash
-        -> Therefore, each empty state trie now points to 0x80, which is not a valid trie node, which crashes @ethereumjs/trie
-      */
-
-      // Setup
-      const stateManager = new DefaultStateManager({ accountCacheOpts })
-      const codeStateManager = new DefaultStateManager({ accountCacheOpts })
-      const address1 = new Address(Buffer.from('a94f5374fce5edbc8e2a8697c15331677e6ebf0b', 'hex'))
-      const account = createAccount()
-      const key1 = Buffer.from('00'.repeat(32), 'hex')
-      const key2 = Buffer.from('00'.repeat(31) + '01', 'hex')
-
-      await stateManager.putAccount(address1, account)
-      await stateManager.putContractStorage(address1, key1, key2)
-      await stateManager.putContractStorage(address1, key2, key2)
-      const root = await stateManager.getStateRoot()
-      // @ts-expect-error
-      const rawNode = await stateManager._trie._db.get(root)
-
-      await codeStateManager.putContractCode(address1, rawNode!)
-
-      let codeSlot1 = await codeStateManager.getContractStorage(address1, key1)
-      let codeSlot2 = await codeStateManager.getContractStorage(address1, key2)
-
-      st.ok(codeSlot1.length === 0, 'slot 0 is empty')
-      st.ok(codeSlot2.length === 0, 'slot 1 is empty')
-
-      const code = await codeStateManager.getContractCode(address1)
-      st.ok(code.length > 0, 'code deposited correctly')
-
-      const slot1 = await stateManager.getContractStorage(address1, key1)
-      const slot2 = await stateManager.getContractStorage(address1, key2)
-
-      st.ok(slot1.length > 0, 'storage key0 deposited correctly')
-      st.ok(slot2.length > 0, 'storage key1 deposited correctly')
-
-      let slotCode = await stateManager.getContractCode(address1)
-      st.ok(slotCode.length === 0, 'code cannot be loaded')
-
-      // Checks by either setting state root to codeHash, or codeHash to stateRoot
-      // The knowledge of the tries should not change
-      let account1 = await stateManager.getAccount(address1)
-      account1!.codeHash = root
-
-      await stateManager.putAccount(address1, account1!)
-
-      slotCode = await stateManager.getContractCode(address1)
-      st.ok(slotCode.length === 0, 'code cannot be loaded') // This test fails if no code prefix is used
-
-      account1 = await codeStateManager.getAccount(address1)
-      account1!.storageRoot = root
-
-      await codeStateManager.putAccount(address1, account1!)
-
-      codeSlot1 = await codeStateManager.getContractStorage(address1, key1)
-      codeSlot2 = await codeStateManager.getContractStorage(address1, key2)
-
-      st.ok(codeSlot1.length === 0, 'slot 0 is empty')
-      st.ok(codeSlot2.length === 0, 'slot 1 is empty')
-
-      st.end()
-    })
-  }
-})
-
-tape('StateManager - Contract code', (tester) => {
-  for (const accountCacheOpts of [{ deactivate: false }, { deactivate: true }]) {
-    const it = tester.test
-    it('should set and get code', async (t) => {
-      const stateManager = new DefaultStateManager({ accountCacheOpts })
-      const address = new Address(Buffer.from('a94f5374fce5edbc8e2a8697c15331677e6ebf0b', 'hex'))
-      const code = Buffer.from(
-        '73095e7baea6a6c7c4c2dfeb977efac326af552d873173095e7baea6a6c7c4c2dfeb977efac326af552d873157',
-        'hex'
-      )
-      const raw = {
-        nonce: '0x0',
-        balance: '0x03e7',
-        stateRoot: '0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421',
-        codeHash: '0xb30fb32201fe0486606ad451e1a61e2ae1748343cd3d411ed992ffcc0774edd4',
-      }
-      const account = Account.fromAccountData(raw)
-      await stateManager.putAccount(address, account)
-      await stateManager.putContractCode(address, code)
-      const codeRetrieved = await stateManager.getContractCode(address)
-      t.ok(code.equals(codeRetrieved))
-      t.end()
-    })
-
-    it('should not get code if is not contract', async (t) => {
-      const stateManager = new DefaultStateManager({ accountCacheOpts })
-      const address = new Address(Buffer.from('a94f5374fce5edbc8e2a8697c15331677e6ebf0b', 'hex'))
-      const raw = {
-        nonce: '0x0',
-        balance: '0x03e7',
-      }
-      const account = Account.fromAccountData(raw)
-      await stateManager.putAccount(address, account)
-      const code = await stateManager.getContractCode(address)
-      t.ok(code.equals(Buffer.alloc(0)))
-      t.end()
-    })
-
-    it('should set empty code', async (t) => {
-      const stateManager = new DefaultStateManager({ accountCacheOpts })
-      const address = new Address(Buffer.from('a94f5374fce5edbc8e2a8697c15331677e6ebf0b', 'hex'))
-      const raw = {
-        nonce: '0x0',
-        balance: '0x03e7',
-      }
-      const account = Account.fromAccountData(raw)
-      const code = Buffer.alloc(0)
-      await stateManager.putAccount(address, account)
-      await stateManager.putContractCode(address, code)
-      const codeRetrieved = await stateManager.getContractCode(address)
-      t.ok(codeRetrieved.equals(Buffer.alloc(0)))
-      t.end()
-    })
-
-    it('should prefix codehashes by default', async (t) => {
-      const stateManager = new DefaultStateManager({ accountCacheOpts })
-      const address = new Address(Buffer.from('a94f5374fce5edbc8e2a8697c15331677e6ebf0b', 'hex'))
-      const code = Buffer.from('80', 'hex')
-      await stateManager.putContractCode(address, code)
-      const codeRetrieved = await stateManager.getContractCode(address)
-      t.ok(codeRetrieved.equals(code))
-      t.end()
-    })
-
-    // TODO: fix this test (likely Jochem)
-    /**it('should not prefix codehashes if prefixCodeHashes = false', async (t) => {
-      const stateManager = new DefaultStateManager({
-        prefixCodeHashes: false,
-      })
-      const address = new Address(Buffer.from('a94f5374fce5edbc8e2a8697c15331677e6ebf0b', 'hex'))
-      const code = Buffer.from('80', 'hex')
-      try {
-        await stateManager.putContractCode(address, code)
-        t.fail('should throw')
-      } catch (e) {
-        t.pass('successfully threw')
-      }
-      t.end()
-    })*/
-  }
-})
-
-tape('StateManager - Contract storage', (tester) => {
-  for (const accountCacheOpts of [{ deactivate: false }, { deactivate: true }]) {
-    const it = tester.test
-
-    it('should throw on storage values larger than 32 bytes', async (t) => {
-      t.plan(1)
-      const stateManager = new DefaultStateManager({ accountCacheOpts })
-      const address = Address.zero()
-      const key = zeros(32)
-      const value = Buffer.from('aa'.repeat(33), 'hex')
-      try {
-        await stateManager.putContractStorage(address, key, value)
-        t.fail('did not throw')
-      } catch (e: any) {
-        t.pass('threw on trying to set storage values larger than 32 bytes')
-      }
-      t.end()
-    })
-
-    it('should strip zeros of storage values', async (t) => {
-      const stateManager = new DefaultStateManager({ accountCacheOpts })
-      const address = Address.zero()
-
-      const key0 = zeros(32)
-      const value0 = Buffer.from('00' + 'aa'.repeat(30), 'hex') // put a value of 31-bytes length with a leading zero byte
-      const expect0 = unpadBuffer(value0)
-      await stateManager.putContractStorage(address, key0, value0)
-      const slot0 = await stateManager.getContractStorage(address, key0)
-      t.ok(slot0.equals(expect0), 'value of 31 bytes padded correctly')
-
-      const key1 = Buffer.concat([zeros(31), Buffer.from('01', 'hex')])
-      const value1 = Buffer.from('0000' + 'aa'.repeat(1), 'hex') // put a value of 1-byte length with two leading zero bytes
-      const expect1 = unpadBuffer(value1)
-      await stateManager.putContractStorage(address, key1, value1)
-      const slot1 = await stateManager.getContractStorage(address, key1)
-
-      t.ok(slot1.equals(expect1), 'value of 1 byte padded correctly')
-      t.end()
-    })
-
-    it('should delete storage values which only consist of zero bytes', async (t) => {
-      const address = Address.zero()
-      const key = zeros(32)
-      const startValue = Buffer.from('01', 'hex')
-
-      const zeroLengths = [0, 1, 31, 32] // checks for arbitrary-length zeros
-      t.plan(zeroLengths.length)
-
-      for (const length of zeroLengths) {
-        const stateManager = new DefaultStateManager({ accountCacheOpts })
-        const value = zeros(length)
-        await stateManager.putContractStorage(address, key, startValue)
-        const currentValue = await stateManager.getContractStorage(address, key)
-        if (!currentValue.equals(startValue)) {
-          // sanity check
-          t.fail('contract value not set correctly')
-        } else {
-          // delete the value
-          await stateManager.putContractStorage(address, key, value)
-          const deleted = await stateManager.getContractStorage(address, key)
-          t.ok(deleted.equals(zeros(0)), 'the storage key should be deleted')
-        }
-      }
-      t.end()
-    })
-
-    it('should not strip trailing zeros', async (t) => {
-      const address = Address.zero()
-      const key = zeros(32)
-      const value = Buffer.from('0000aabb00', 'hex')
-      const expect = Buffer.from('aabb00', 'hex')
-      const stateManager = new DefaultStateManager({ accountCacheOpts })
-      await stateManager.putContractStorage(address, key, value)
-      const contractValue = await stateManager.getContractStorage(address, key)
-      t.ok(contractValue.equals(expect), 'trailing zeros are not stripped')
-      t.end()
     })
   }
 })
